requirements.txt
Django==2.0.5
django-suit==0.2.26
django-suit-redactor==0.0.4
pytz==2018.4
